+++ src/client/java/jaggwagg/frozen_apocalypse/client/FrozenApocalypseClient.java
@@ -3,6 +3,7 @@ package jaggwagg.frozen_apocalypse.client;
 import jaggwagg.frozen_apocalypse.FrozenApocalypse;
 import jaggwagg.frozen_apocalypse.client.network.ModClientRecievers;
 import jaggwagg.frozen_apocalypse.client.registry.ModBlockRenderLayerMaps;
+import jaggwagg.frozen_apocalypse.client.registry.ModEntityRenderers;
 import net.fabricmc.api.ClientModInitializer;
 
 public class FrozenApocalypseClient implements ClientModInitializer {
@@ -10,8 +11,9 @@ public class FrozenApocalypseClient implements ClientModInitializer {
 
     @Override
     public void onInitializeClient() {
-        ModClientRecievers.init();
         ModBlockRenderLayerMaps.init();
+        ModClientRecievers.init();
+        ModEntityRenderers.init();
 
         FrozenApocalypse.LOGGER.info(FrozenApocalypse.MOD_ID + ": Initialized client successfully");
     }