+++ src/main/java/jaggwagg/frozen_apocalypse/entity/mob/CryoboomerEntity.java
@@ -1,0 +1,11 @@
+package jaggwagg.frozen_apocalypse.entity.mob;
+
+import net.minecraft.entity.EntityType;
+import net.minecraft.entity.mob.CreeperEntity;
+import net.minecraft.world.World;
+
+public class CryoboomerEntity extends CreeperEntity {
+    public CryoboomerEntity(EntityType<? extends CryoboomerEntity> entityType, World world) {
+        super(entityType, world);
+    }
+}