+++ src/main/java/jaggwagg/frozen_apocalypse/entity/mob/FrostbiteEntity.java
@@ -1,0 +1,27 @@
+package jaggwagg.frozen_apocalypse.entity.mob;
+
+import net.minecraft.entity.Entity;
+import net.minecraft.entity.EntityType;
+import net.minecraft.entity.LivingEntity;
+import net.minecraft.entity.effect.StatusEffectInstance;
+import net.minecraft.entity.effect.StatusEffects;
+import net.minecraft.entity.mob.ZombieEntity;
+import net.minecraft.world.World;
+
+public class FrostbiteEntity extends ZombieEntity {
+    public FrostbiteEntity(EntityType<? extends FrostbiteEntity> entityType, World world) {
+        super(entityType, world);
+    }
+
+    @Override
+    public boolean tryAttack(Entity target) {
+        boolean attacked = super.tryAttack(target);
+
+        if (attacked && this.getMainHandStack().isEmpty() && target instanceof LivingEntity) {
+            int duration = (int) this.getWorld().getLocalDifficulty(this.getBlockPos()).getLocalDifficulty();
+            ((LivingEntity)target).addStatusEffect(new StatusEffectInstance(StatusEffects.SLOWNESS, 140 * duration), this);
+        }
+
+        return attacked;
+    }
+}