+++ src/main/java/jaggwagg/frozen_apocalypse/mixin/entity/mob/HostileEntityMixin.java
@@ -1,6 +1,8 @@
 package jaggwagg.frozen_apocalypse.mixin.entity.mob;
 
 import jaggwagg.frozen_apocalypse.FrozenApocalypse;
+import jaggwagg.frozen_apocalypse.registry.ModEntityTypes;
+import net.minecraft.entity.EntityType;
 import net.minecraft.entity.mob.HostileEntity;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.random.Random;
@@ -10,6 +12,7 @@ import net.minecraft.world.WorldView;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(HostileEntity.class)
@@ -29,6 +32,31 @@ public abstract class HostileEntityMixin {
             if (world.getLightLevel(LightType.SKY, pos) > 0) {
                 cir.setReturnValue(1.0f);
             }
+        }
+    }
+
+    @Inject(method = "tickMovement", at = @At("HEAD"))
+    private void tickMovement(CallbackInfo ci) {
+        HostileEntity hostileEntity = (HostileEntity) (Object) this;
+
+        if (hostileEntity.getWorld().isClient || !hostileEntity.isAlive() || hostileEntity.isAiDisabled()) {
+            return;
+        }
+
+        if (!FrozenApocalypse.apocalypseLevel.canFreezeEntities() || hostileEntity.getBlockPos().getY() < FrozenApocalypse.apocalypseLevel.getFreezingYLevel()) {
+            return;
+        }
+
+        if (hostileEntity.getType().equals(EntityType.CREEPER)) {
+            hostileEntity.convertTo(ModEntityTypes.RegisteredMobEntityTypes.CRYOBOOMER.getEntityType(), true);
+        } else if (hostileEntity.getType().equals(EntityType.ENDERMAN)) {
+            hostileEntity.convertTo(ModEntityTypes.RegisteredMobEntityTypes.SHIVERSTARE.getEntityType(), true);
+        } else if (hostileEntity.getType().equals(EntityType.SKELETON)) {
+            hostileEntity.convertTo(EntityType.STRAY, true);
+        } else if (hostileEntity.getType().equals(EntityType.SPIDER)) {
+            hostileEntity.convertTo(ModEntityTypes.RegisteredMobEntityTypes.ICEWEAVER.getEntityType(), true);
+        } else if (hostileEntity.getType().equals(EntityType.ZOMBIE)) {
+            hostileEntity.convertTo(ModEntityTypes.RegisteredMobEntityTypes.FROSTBITE.getEntityType(), true);
         }
     }
 }