+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ##### Added
 - Added a 9th apocalypse level, sun is now 90% smaller, entities freeze very quickly
+- Added 4 new mob variants: Cryoboomer (creeper), Frostbite (zombie), Iceweaver (spider), Shiverstare (enderman)
+- These variants are created if the corresponding mob is above the freezing Y level
 - Added 4 new blocks: frosted grass, dead grass, dead leaves, and icicles
 - Each new block spawns during certain apocalypse levels
 - Added config option for minimum light level to avoid freezing for each apocalypse level, further levels increase the minimum light level
@@ -23,6 +25,7 @@
 ##### Fixed
 - Apocalypse now lags the world even less due to multiple optimizations and feature removal
 - Sun size change packet being registered every tick and returning incorrect data type
+- Fixed a bunch of leaves remaining after decaying on trees (still might be some though)
 
 ##### Removed
 - Removed the ability to add custom light source check (contributed to significant performance issues)